Integer totals (#25)
* added test to fail when posting a float total

* receipt model and schema now have integer total instead of numeric/float

* fixed crud tests to use integers instead of floats

* added crud test to fail on float total

* fixed api tests to use ints

* removed unnecessary pytest.approx from receipt tests

* fixed receipt test data to have int total

* improved receiptlines tests to avoid pytest.approx precision pitfall

* created DB migration

* added response models to routers

* added missing Optional type from name

* added missing orm_mode from chain and paymentmethod schemas

* use different DB based on env var mode, for development purposes

* utility script for running API in dev mode

* script to fetch API's openapi schema

* get receipt_api_schema.json

* test for schema being up-to-date

* model and test for ReceiptCreate

* more comprehensive module docstring

* added class to interact with API

* load.py now uses ReceiptAPI and ReceipCreate

* added pytest-mock

* added test data

* PUT /receipts path now has receipt_id, as it should

* test for api wrapper, refactoring

* transformations now return integer totals for receipts

+++ backend/app/models/receipt.py
@@ -1,5 +1,5 @@
 from sqlalchemy.sql.schema import Column, ForeignKey
-from sqlalchemy.sql.sqltypes import TIMESTAMP, Numeric, Text
+from sqlalchemy.sql.sqltypes import TIMESTAMP, Integer, Text
 
 from backend.app.db.base_class import Base
 
@@ -9,7 +9,7 @@ class Receipt(Base):
 
     id = Column(Text, primary_key=True)
     reprint = Column(Text)
-    total = Column(Numeric)
+    total = Column(Integer)
     etag = Column(Text)
     datetime = Column(TIMESTAMP(timezone=True))
     store_id = Column(Text, ForeignKey('store.id'))
